web-worker/requirements.txt
boto3
requests
aiohttp
python-dotenv
black==24.3.0
flake8==7.0.0
